Fix some PEP-8. Bump version number. Add gitignore for IDEA.
